02-llm-client/requirements.txt
mcp
websockets>=11.0.0
anyio>=3.7.0
python-dotenv
sseclient-py
requests
openai>=1.0.0
